webnn/exp.html

<!doctype html>
<meta charset=utf-8>
<title>test element-wise unary operations</title>
<link rel="help" href="https://webmachinelearning.github.io/webnn/#api-mlgraphbuilder-unary">
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>
<script src="./webnn-polyfill.js"></script>
<div id=log></div>
<script type="module">
  'use strict';
  import * as utils from './utils.js';
  let builder;
  setup(() => {
    const context = navigator.ml.createContext();
    builder = new MLGraphBuilder(context);
  })

  function testUnary(op, input, expected, shape) {
    const x = builder.input('x', {type: 'float32', dimensions: shape});
    const y = builder[op](x);
    const graph = builder.build({y});
    const inputs = {'x': new Float32Array(input)};
    const outputs = {'y': new Float32Array(utils.sizeOfShape(shape))};
    graph.compute(inputs, outputs);
    // assert_array_approx_equals_ulp(op, outputs.y, expected, 0);
    for (let i = 0; i < outputs.y.length; i++) {
      assert_approx_equals( outputs.y[i], expected[i], 1.0/1024, "test exp");
    }
  }


  test(() => {
    testUnary(
        'exp',
        [
        -22.99204874532912,
        -20.743229075086745,
        -18.993467419350843,
        -7.540661071653204,
        -2.543807029272031,
        -8.021520259264364,
        -2.9926157522384074,
        -8.468723600914828,
        -12.788521558154791,
        -7.444099880551349,
        -16.59915103066328,
        -3.7920975843961244,
        -6.553587659308249,
        -13.991944399901854,
        -14.272076274218385,
        -4.491299602004425,
        -3.365216421474159,
        -21.23593355569512,
        -2.922388619369329,
        -9.674529849030655,
        -15.327549466549947,
        -3.3039481280052527,
        -31.825224375967387,
        -13.792974631069725,
        -0.6537282725784639,
        -0.369773078862929,
        -0.6310476476790319,
        -0.2339118443287478,
        -0.31822392861546467,
        -0.1455715074310142,
        -0.8668789462404793,
        -0.8965133564460195,
        -0.3338304074128866,
        -0.4252149891957875,
        -0.6670083494065631,
        -0.9138738140030531,
        -0.3686426497001547,
        -0.7885576576698421,
        -0.938683915789472,
        -0.0945066370240284,
        -0.6534228693832069,
        -0.6278267054843902,
        -0.20885064570898626,
        -0.7632029460209122,
        -0.3146389306533488,
        -0.6757310922577509,
        -0.5069854844935684,
        -0.23438132610503826,
        20.813061167070877,
        30.910333904837067,
        4.962480403161663,
        20.0467240906249,
        4.694365060121655,
        22.620644348144452,
        6.992251016749599,
        1.2960263691619929,
        21.558996466377643,
        6.216680878598737,
        6.343355886430736,
        24.878306448588418,
        5.480594134063608,
        1.84147599931147,
        14.240338631858862,
        24.46251196998665,
        10.04045060355589,
        6.695826149584278,
        16.158195085828005,
        5.814379130575155,
        11.261562196063316,
        3.6164375614735746,
        9.914684439586907,
        9.314918894083625,
        ],
        [
          1.0343799702128246e-10, 
9.802347078951733e-10, 
5.639516966448492e-9, 
0.0005310464459828795, 
0.07856672389748613, 
0.00032832051091804783, 
0.05015606913585237, 
0.00020993269031914735, 
0.0000027926445096052763, 
0.0005848823274924007, 
6.181306848600785e-8, 
0.02254825534326579, 
0.0014249940234959484, 
8.382542345967282e-7, 
6.334553797109671e-7, 
0.0112060709099391, 
0.03455453696487981, 
5.988957603797542e-10, 
0.053805013987173966, 
0.000062864442948692, 
2.204600215745022e-7, 
0.036737835019307594, 
1.508275470953635e-14, 
0.0000010227917792367717, 
0.5201030715244482, 
0.6908910906424834, 
0.5320341246178955, 
0.7914315820027613, 
0.7274398755706821, 
0.8645280676824401, 
0.42026116217858056, 
0.4079896973679189, 
0.7161752299386734, 
0.6536292467971649, 
0.5132417232312166, 
0.400967936388636, 
0.6916725356805613, 
0.4544998671472885, 
0.39114227293215936, 
0.9098216960975325, 
0.52026193692212, 
0.5337505385299562, 
0.8115164300537903, 
0.4661709129922622, 
0.7300524262384879, 
0.5087843163283117, 
0.6023085132327102, 
0.7910601065051417, 
1093950362.1065366, 
26557515807997.72, 
142.94792500908824, 
508372035.3858906, 
109.32936902288711, 
6668390398.21978, 
1088.1682061961842, 
3.654745167518796, 
2306498628.5912433, 
501.0374649351999, 
568.701610627777, 
63754553672.34833, 
239.989250813914, 
6.305838809277489, 
1529327.465744619, 
42066265840.60161, 
22935.715503452648, 
809.0220314133451, 
10409142.722216586, 
335.0832909338617, 
77773.97915518607, 
37.20479167581219, 
20225.196136128863, 
11102.42569029876,
      ],
      [6,3,4]);
  }, 'exp');

</script>
